Refactor blog test file structure to closer map to source structure
